feat: move thread to reactive state fully
