+++ package/src/components/Channel/hooks/useCreateMessagesContext.ts
@@ -6,8 +6,6 @@ export const useCreateMessagesContext = ({
   additionalPressableProps,
   channelId,
   customMessageSwipeAction,
-  deleteMessage,
-  deleteReaction,
   disableTypingIndicator,
   dismissKeyboardOnMessageTouch,
   enableMessageGroupingByUser,
@@ -46,13 +44,9 @@ export const useCreateMessagesContext = ({
   onPressMessage,
   reactionListPosition,
   reactionListType,
-  removeMessage,
-  retrySendMessage,
   selectReaction,
-  sendReaction,
   shouldShowUnreadUnderlay,
   supportedReactions,
-  updateMessage,
   urlPreviewType,
 }: MessagesContextValue & {
   /**
@@ -69,8 +63,6 @@ export const useCreateMessagesContext = ({
     () => ({
       additionalPressableProps,
       customMessageSwipeAction,
-      deleteMessage,
-      deleteReaction,
       disableTypingIndicator,
       dismissKeyboardOnMessageTouch,
       enableMessageGroupingByUser,
@@ -109,13 +101,9 @@ export const useCreateMessagesContext = ({
       onPressMessage,
       reactionListPosition,
       reactionListType,
-      removeMessage,
-      retrySendMessage,
       selectReaction,
-      sendReaction,
       shouldShowUnreadUnderlay,
       supportedReactions,
-      updateMessage,
       urlPreviewType,
     }),
     // eslint-disable-next-line react-hooks/exhaustive-deps
